Move GenParamGetter
